--- examples/external-audio/index.html
@@ -1,81 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-US" translate="no">
-	<head>
-		<meta charset="utf-8">
-		<meta http-equiv="X-UA-Compatible" content="IE=edge">
-		<meta name="viewport" content="width=device-width, initial-scale=1.0">
-		<title>External audio track example | Mediabunny</title>
-		<meta name="description" content="Select or drop a video file, and Mediabunny will attach a synthesized voiceover-like audio track to it without taking ownership of the output file.">
-		<script type="module" src="../base.ts"></script>
-		<script type="module" src="./external-audio.ts"></script>
-		<link rel="stylesheet" href="../base.css">
-		<link rel="icon" href="../../docs/public/mediabunny-logo.svg">
-		<link rel="canonical" href="https://mediabunny.dev/examples/external-audio/">
-		<meta property="og:site_name" content="Mediabunny">
-		<meta property="og:type" content="website">
-		<meta property="og:title" content="External audio track example | Mediabunny">
-		<meta property="og:description" content="Select or drop a video file, and Mediabunny will attach a synthesized voiceover-like audio track to it without taking ownership of the output file.">
-		<meta property="og:url" content="https://mediabunny.dev/examples/external-audio/">
-		<meta property="og:image" content="https://mediabunny.dev/mediabunny-og-image.png">
-		<meta property="og:locale" content="en-US">
-		<meta name="twitter:card" content="summary_large_image">
-		<meta name="twitter:site" content="@vanilagy">
-		<meta name="twitter:title" content="External audio track example | Mediabunny">
-		<meta name="twitter:description" content="Select or drop a video file, and Mediabunny will attach a synthesized voiceover-like audio track to it without taking ownership of the output file.">
-		<meta name="twitter:image" content="https://mediabunny.dev/mediabunny-og-image.png">
-		<script type="application/ld+json">{"@context":"https://schema.org","@type":"BreadcrumbList","itemListElement":[{"@type":"ListItem","position":1,"name":"Examples","item":"https://mediabunny.dev/examples"},{"@type":"ListItem","position":2,"name":"External audio track"}]}</script>
-	</head>
-
-	<body class="flex flex-col items-center py-10 bg-zinc-50 text-zinc-800 dark:bg-zinc-900 dark:text-zinc-200 px-2">
-		<h1 class="text-3xl font-bold text-violet-500 text-center">External audio track example</h1>
-		<p class="max-w-lg text-center">
-			Select or drop a video file (with or without audio), and Mediabunny will synthesize a short
-			voiceover-like audio track and attach it to the video. The conversion doesn't own the output file here —
-			we do, and we add the new audio track ourselves.
-		</p>
-
-		<div class="flex flex-col items-center gap-1">
-			<div class="flex gap-2 mt-4">
-				<button id="select-file" class="rounded-lg bg-zinc-200 dark:bg-zinc-750 hover:bg-zinc-300 dark:hover:bg-zinc-700 px-5 py-2">
-					Select local file
-				</button>
-
-				<button id="load-url" class="rounded-lg bg-zinc-200 dark:bg-zinc-750 hover:bg-zinc-300 dark:hover:bg-zinc-700 px-5 py-2">
-					Load remote URL
-				</button>
-			</div>
-
-			<a id="sample-file-download" download="big-buck-bunny-trimmed.mp4" class="hover:underline text-xs opacity-50 hover:opacity-70">
-				Download sample file
-			</a>
-		</div>
-
-		<p class="text-xs opacity-60 mt-2 mx-auto text-center max-w-[36rem]" id="file-name"></p>
-
-		<hr class="w-full max-w-96 my-4 border-zinc-300 dark:border-zinc-700" style="display: none;">
-
-		<p id="error-element" class="text-red-500 mx-auto text-center max-w-[36rem]"></p>
-
-		<div class="w-full max-w-80 h-2 rounded-full bg-zinc-200 dark:bg-zinc-750 overflow-hidden" id="progress-bar-container" style="display: none;">
-			<div class="h-full bg-violet-500 w-0" id="progress-bar"></div>
-		</div>
-		<p class="text-xs font-medium mt-1" id="status-text" style="display: none;"></p>
-
-		<video controls class="rounded-md mt-4" style="display: none;"></video>
-		<p class="text-xs font-medium mt-1" id="result-info" style="display: none;"></p>
-
-		<a href="/" class="fixed top-0 left-0 flex gap-2 py-2 px-5 items-center">
-			<img src="../../docs/public/mediabunny-logo.svg" class="size-6">
-			<p class="text-sm font-medium">Mediabunny</p>
-		</a>
-
-		<a
-			href="https://github.com/Vanilagy/mediabunny/tree/main/examples/external-audio"
-			target="_blank"
-			class="flex items-center gap-2 fixed top-0 right-0 py-2 px-5 bg-zinc-200 dark:bg-zinc-750 hover:bg-zinc-300 dark:hover:bg-zinc-700 rounded-bl-xl"
-		>
-			<img src="../../docs/assets/github-mark.svg" class="size-6 dark:invert">
-			<p>View source code</p>
-		</a>
-	</body>
-</html>
